refactor structures serializers
