--- nanome/_internal/macro/serialization/__init__.py
@@ -1,2 +1,0 @@
-from . import *
-from .macro_serializer import _MacroSerializer
