--- nanome/_internal/structure/serialization/chain_serializer.py
@@ -1,35 +1,0 @@
-from nanome._internal.util.type_serializers import ArraySerializer, StringSerializer
-from . import _ResidueSerializer
-from .. import _Chain
-
-from nanome._internal.util.type_serializers import TypeSerializer
-
-
-class _ChainSerializer(TypeSerializer):
-    def __init__(self, shallow=False):
-        self.shallow = shallow
-        self.array_serializer = ArraySerializer()
-        self.array_serializer.set_type(_ResidueSerializer())
-        self.string = StringSerializer()
-
-    def version(self):
-        return 0
-
-    def name(self):
-        return "Chain"
-
-    def serialize(self, version, value, context):
-        context.write_long(value._index)
-        if (self.shallow):
-            context.write_using_serializer(self.array_serializer, [])
-        else:
-            context.write_using_serializer(self.array_serializer, value._residues)
-        context.write_using_serializer(self.string, value._name)
-
-    def deserialize(self, version, context):
-        chain = _Chain._create()
-        chain._index = context.read_long()
-
-        chain._set_residues(context.read_using_serializer(self.array_serializer))
-        chain._name = context.read_using_serializer(self.string)
-        return chain
